+++ inmoose/utils/__init__.py
@@ -1,2 +1,2 @@
 from .factor import Factor, gl, asfactor
-from .stats import rnbinom
+from .stats import rnbinom, dnbinom_mu, dnorm, pnorm, pt